Update tests to use Intern
Fixes #153
