+++ .travis.yml
@@ -1,0 +1,12 @@
+sudo: false
+language: node_js
+node_js:
+  - "0.10"
+cache:
+directories:
+  - node_modules
+install:
+  - "travis_retry npm install grunt-cli"
+  - "travis_retry npm install"
+script:
+  - "grunt ci"
