--- test/karma-setup.js
@@ -1,5 +1,0 @@
-mocha.setup({
-  ui: 'tdd',
-  slow: 1000,
-  timeout: 30000
-});
