--- test/runner.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<title>PointerEvent tests</title>
-<meta charset="UTF-8">
-<!-- Mocha -->
-<link rel="stylesheet" href="../node_modules/mocha/mocha.css">
-<script src="../node_modules/mocha/mocha.js"></script>
-<!-- Chai -->
-<script src="../node_modules/chai/chai.js"></script>
-<!-- Chai Spies -->
-<script src="../node_modules/chai-spies/chai-spies.js"></script>
-<!-- PointerEvents -->
-<script src="../dist/PEP.js"></script>
-<!-- Test Setup -->
-<div id="mocha"></div>
-<script src="karma-setup.js"></script>
-<!-- Tests -->
-<script src="setup.js"></script>
-<script src="loader.js"></script>
-<script src="constructor.js"></script>
-<script src="dispatcher.js"></script>
-<script src="pointermap.js"></script>
-<script src="capture.js"></script>
-<!-- Run Mocha -->
-<script>
-  mocha.run();
-</script>
