+++ tests/intern-local.js
@@ -1,0 +1,5 @@
+define([ './intern' ], function (config) {
+	config.environments = [ { browserName: 'chrome' } ];
+	config.tunnel = 'NullTunnel';
+	return config;
+});
